# Tradgardsgolf.ApiClient/Tradgardsgolf API/Models/CreateLoginModel.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace Tradgardsgolf.ApiClient.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    public partial class CreateLoginModel
    {
        /// <summary>
        /// Initializes a new instance of the CreateLoginModel class.
        /// </summary>
        public CreateLoginModel() { }

        /// <summary>
        /// Initializes a new instance of the CreateLoginModel class.
        /// </summary>
        public CreateLoginModel(string email, string password)
        {
            Email = email;
            Password = password;
        }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "email")]
        public string Email { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "password")]
        public string Password { get; set; }

        /// <summary>
        /// Validate the object. Throws ValidationException if validation fails.
        /// </summary>
        public virtual void Validate()
        {
            if (Email == null)
            {
                throw new ValidationException(ValidationRules.CannotBeNull, "Email");
            }
            if (Password == null)
            {
                throw new ValidationException(ValidationRules.CannotBeNull, "Password");
            }
        }
    }
}