# Tradgardsgolf.ApiClient/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tradgardsgolf API",
    "version": "v1"
  },
  "paths": {
    "/api/Authentication": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "Authenticate",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CredentialsModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthenticationResponse"
            }
          }
        }
      }
    },
    "/api/Authorization/IsAuthorized": {
      "get": {
        "tags": [
          "Authorization"
        ],
        "operationId": "IsAuthorized",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Login/Create": {
      "post": {
        "tags": [
          "Login"
        ],
        "operationId": "CreateLogin",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateLoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CredentialsModel": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "AuthenticationResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string",
          "readOnly": true
        },
        "token": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "CreateLoginModel": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. <br /> <br /> \r\n                      Enter 'Bearer' [space] and then your token in the text input below. <br /> <br /> \r\n                      Example: 'Bearer 12345abcdef'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}